requirements.txt
discord.py==1.5.1
HyruleCompendium==0.0.2